+++ tests/integration/chart_table01.rs
@@ -6,10 +6,10 @@
 // Copyright 2022-2023, John McNamara, jmcnamara@cpan.org
 
 use crate::common;
-use rust_xlsxwriter::{Chart, ChartSeries, ChartTable, ChartType, Workbook, XlsxError};
+use rust_xlsxwriter::{Chart, ChartDataTable, ChartSeries, ChartType, Workbook, XlsxError};
 
 // Create rust_xlsxwriter file to compare against Excel file.
-fn create_new_xlsx_file(filename: &str) -> Result<(), XlsxError> {
+fn create_new_xlsx_file_1(filename: &str) -> Result<(), XlsxError> {
     let mut workbook = Workbook::new();
 
     let worksheet = workbook.add_worksheet();
@@ -29,9 +29,45 @@ fn create_new_xlsx_file(filename: &str) -> Result<(), XlsxError> {
         .push_series(ChartSeries::new().set_values(("Sheet1", 0, 1, 4, 1)))
         .push_series(ChartSeries::new().set_values(("Sheet1", 0, 2, 4, 2)));
 
-    let table = ChartTable::default();
+    let table = ChartDataTable::default();
 
-    chart.set_table(&table);
+    chart.set_data_table(&table);
+
+    worksheet.insert_chart(8, 4, &chart)?;
+
+    workbook.save(filename)?;
+
+    Ok(())
+}
+
+// Use explicit default values.
+fn create_new_xlsx_file_2(filename: &str) -> Result<(), XlsxError> {
+    let mut workbook = Workbook::new();
+
+    let worksheet = workbook.add_worksheet();
+
+    // Add some test data for the chart(s).
+    let data = [[1, 2, 3], [2, 4, 6], [3, 6, 9], [4, 8, 12], [5, 10, 15]];
+    for (row_num, row_data) in data.iter().enumerate() {
+        for (col_num, col_data) in row_data.iter().enumerate() {
+            worksheet.write_number(row_num as u32, col_num as u16, *col_data)?;
+        }
+    }
+
+    let mut chart = Chart::new(ChartType::Column);
+    chart.set_axis_ids(61355520, 61357056);
+    chart
+        .push_series(ChartSeries::new().set_values(("Sheet1", 0, 0, 4, 0)))
+        .push_series(ChartSeries::new().set_values(("Sheet1", 0, 1, 4, 1)))
+        .push_series(ChartSeries::new().set_values(("Sheet1", 0, 2, 4, 2)));
+
+    let table = ChartDataTable::new()
+        .show_horizontal_borders(true)
+        .show_vertical_borders(true)
+        .show_outline_borders(true)
+        .show_legend_keys(false);
+
+    chart.set_data_table(&table);
 
     worksheet.insert_chart(8, 4, &chart)?;
 
@@ -41,10 +77,23 @@ fn create_new_xlsx_file(filename: &str) -> Result<(), XlsxError> {
 }
 
 #[test]
-fn test_chart_table01() {
+fn test_chart_table01_1() {
     let test_runner = common::TestRunner::new()
         .set_name("chart_table01")
-        .set_function(create_new_xlsx_file)
+        .set_function(create_new_xlsx_file_1)
+        .unique("1")
+        .initialize();
+
+    test_runner.assert_eq();
+    test_runner.cleanup();
+}
+
+#[test]
+fn test_chart_table01_2() {
+    let test_runner = common::TestRunner::new()
+        .set_name("chart_table01")
+        .set_function(create_new_xlsx_file_2)
+        .unique("2")
         .initialize();
 
     test_runner.assert_eq();